% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acgEllipse.R
\name{acgEllipse}
\alias{acgEllipse}
\title{Graphical Representations of Angular Central Gaussian Distributions from qPLM}
\usage{
acgEllipse(angGaussObj)
}
\arguments{
\item{angGaussObj}{Output from the \code{\link{angGaussSumm}} function.}
}
\value{
Silent. Called for the side-effect of producing an \code{rgl} object.
}
\description{
\code{acgEllipse} creates an interactive three-dimensional figure
  portraying the distribution of axial orientation data.
}
\details{
Uses the \code{ellipsoid3d()} function from the \code{shapes3d} demo
  in package \code{rgl} to plot the eigenvalues and eigenvectors of the
  lambda matrix, describing the major axes of an Angular Central Gaussian
  distribution.
}
\references{
Tyler, D.E., 1987. Statistical analysis for the angular central
  Gaussian distribution on the sphere. \emph{Biometrika, 74(3)}: 579 - 589.
}
\seealso{
Other qPLM Illustration Functions: \code{\link{qPLMKeymaker}},
  \code{\link{qPLMSphereSkinner}}, \code{\link{qPLMTiff}}
}
