% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angGaussSum.R
\name{angGaussSumm}
\alias{angGaussSumm}
\title{Estimate and Summarize Angular Central Gaussian Distribution for Axial
  Data}
\usage{
angGaussSumm(X, alpha = 0.05, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{X}{An \code{[n,3]} matrix of Euclidean coordinates for axial
orientations, for general use. \code{angGaussSumm} will also parse the
orientations from a \code{qPLMtab} object referenced by this argument,
without requiring specific column addresses.}

\item{alpha}{Alpha value corresponding to confidence interval \code{1 - CI}.
Default is 95\% CI.}

\item{tol}{Tolerance for convergence of iterative algorithm. Default value is
the default for the \code{all.equal()} function. Increase this value if the
function returns an identity matrix for the lambda estimate.}
}
\value{
A list with four components: \enumerate{ \item $lambda.matrix: a
  matrix of lambda values, \item $lambda.eigval: the eigenvalues of the
  lambda matrix, \item $lambda.eigvec: the eigenvectors of the lambda matrix
  (major orientations of the ACG distribution), and \item $aCI: angular
  confidence intervals for first axis in second axis plane (I-II), first axis
  in third axis plane (I-III), and second axis in third axis plane (II-III),
  given in both radians and degrees.}
}
\description{
\code{angGaussSumm} estimates the angular central Gaussian
  distribution of 3D axial data using Tyler's (1987) maximum likelihood
  method.
}
\examples{
#oldwd<-getwd()
#setwd(system.file("extdata", package = "microTransit"))
#load("testqPLMtab.R")
testqPLMGauss_tab<-angGaussSumm(testqPLMtab)
save(testqPLMGauss_tab, file = "testqPLMGauss_tab.R")
#setwd(oldwd)

}
\references{
Tyler, D.E., 1987. Statistical analysis for the angular central
  Gaussian distribution on the sphere. \emph{Biometrika, 74(3)}: 579 - 589.
}
\seealso{
Other qPLM Analysis Functions: 
\code{\link{centroidCorr}()},
\code{\link{pullCluster}()},
\code{\link{pullROI}()},
\code{\link{pullqPLMSample}()},
\code{\link{qPLMClust}()},
\code{\link{qPLMTabulate}()}
}
\concept{qPLM Analysis Functions}
