% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroidCorr.R
\name{centroidCorr}
\alias{centroidCorr}
\title{Set Tangential Frame of Reference for Cortical Bone Axial Data}
\usage{
centroidCorr(qPLMtab)
}
\arguments{
\item{qPLMtab}{qPLMtab object to process.}
}
\value{
A \code{qPLMtab} object; a matrix with columns \enumerate{ \item (1)
  theta, \item (2) 'corrected' phi, \item (3-5) the 'unwrapped' Euclidean x,
  y, and z axis coordinates, and \item (6-7) the original pixel u,v
  positions.}
}
\description{
\code{centroidCorr} changes the in-plane orientation data of a
  \code{qPLMtab} object from global reference frame to orientation relative
  to a tangent from the centroid of the object cross-section. Resulting
  orientations are in a sense 'unwrapping' the cortex around the centroid.
}
\details{
For some questions (e.g., resistance of cortical bone to torsion),
  orientation in plane relative to a centroid is informative. This function
  returns orientation relative to a 'naive' assumption of the structural
  centroid of a cross-section.
}
\examples{
#oldwd<-getwd()
#setwd(system.file("extdata", package = "microTransit"))
#load("testqPLMtab.R")
#setwd(oldwd)

testCorrtab<-centroidCorr(testqPLMtab)

}
\seealso{
Other qPLM Analysis Functions: 
\code{\link{angGaussSumm}()},
\code{\link{pullCluster}()},
\code{\link{pullROI}()},
\code{\link{pullqPLMSample}()},
\code{\link{qPLMClust}()},
\code{\link{qPLMTabulate}()}
}
\concept{qPLM Analysis Functions}
