% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jonesRet.R
\name{jonesRet}
\alias{jonesRet}
\title{Jones Matrix Representation of a Retarder Element}
\usage{
jonesRet(retardance = NULL, biref, theta.deg, d.um, lambda, slow.phi.deg)
}
\arguments{
\item{retardance}{Shortcut to produce a specified retardance. Default is
null.}

\item{biref}{Birefringence of the material to be modeled. Required if
\code{retardance} is null.}

\item{theta.deg}{Out-of-plane orientation of the slow axis of the material.
Required if \code{retardance} is null.}

\item{d.um}{Thickness of the element in micrometers. Required if
\code{retardance} is null.}

\item{lambda}{Wavelength for simulation. Required if \code{retardance} is
null.}

\item{slow.phi.deg}{In-plane orientation of the slow axis of the material.}
}
\value{
A Jones matrix.
}
\description{
\code{jonesRet} creates a Jones matrix for a retarder element,
  either with \enumerate{ \item (a) specified retardance or \item (b)
  specified birefringence, orientation, thickness, and wavelength,} at
  specified slow axis orientation.
}
\details{
The Jones calculus provides a means to represent polarization of light with matrix and vector elements (Fowles 1989).
}
\references{
Fowles, G. R. (1989). Introduction to modern optics. Courier Corporation.
}
\seealso{
Other qPLM Simulation Functions: 
\code{\link{Rotosim}()},
\code{\link{jonesPol}()}
}
\concept{qPLM Simulation Functions}
