% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KaminskyqPLM.R
\name{KaminskyqPLM}
\alias{KaminskyqPLM}
\title{Interactively Build qPLMarr Object From Werner Kaminsky's Rotopol(tm)
  Format Images}
\usage{
KaminskyqPLM(sample.name, north.thickness, south.thickness = NULL,
  west.thickness = NULL, east.thickness = NULL, wavelength = 532,
  birefringence = 0.005, pixel = 7.5832259, up = 0, mask = TRUE,
  data.type = "generic")
}
\arguments{
\item{sample.name}{A name that will be used as a stem to label output files
from this function (images, etc.).}

\item{north.thickness}{Numerical; the thickness of the (masked) specimen at
  its "North" (top center) edge. If no mask is specified, this thickness
  should be measured at the top center edge of the image.

  N, S, W, & E thicknesses can be specified separately to model 'wedged'
  specimens. If the specimen can be assumed to have a constant thickness
  across the field of view, the N value can be measured anywhere on the
  image, and will be used to fill in S, W, & E by default.}

\item{south.thickness, west.thickness, east.thickness}{Numerical; optional
thickness measurements from each corner of the}

\item{wavelength}{Numerical; the center of the illumination wavelength.
Default is a 532nm green filter.}

\item{birefringence}{Numerical; the expected birefringence of the specimen.
Default value is a semi-empirical lab standard of 0.0005 for bone. Images
with several materials of varying birefringence currently require multiple
input runs with separate masks for each material.}

\item{pixel}{Numerical; the absolute size of a pixel in the specimen plane.
Intended to be in microns, but the value is arbitrary and not currently
linked to specific dimensions in other functions. Default value is lab
standard for a Leica A6 macroscope at 1.25x.}

\item{up}{Numerical; a 'map-view' compass orientation to track a specific
axis. Can be used, for example, to specify the dorsal side of a long bone
cross-section. This value is passed to the qPLMarr object's attributes and
does not influence the values in the result array.}

\item{mask}{If \code{mask = TRUE}, \code{KaminskyqPLM} will prompt the user
to select a binary mask image with the same dimensions as the input
bitmaps. White pixels in the mask represent pixels from the input bitmaps
that will be included in the qPLM output; black pixels are excluded.}

\item{data.type}{Character string; description of section type in broad
terms, e.g., <<"diaphyseal cross section">>. Keeps specimen data with
qPLMarr object as an attribute, and can also be used for watchdogs in
analysis functions. For example, applying \code{centroidCorr} to an object
will add <<"centroid-corrected">> to the data.type attribute. In future
versions, \code{centroidCorr} may prompt the user for confirmation if it is
applied to an object that is already labeled as centroid-corrected. Default
value is "generic."}
}
\value{
A qPLMarr object that records pixel-by-pixel estimates of I
  (transmissivity), phi (angular orientation in plane), and theta (angular
  orientation out of plane). qPLMarr objects also keep relevant data
  regarding the specimen: thickness, illumination wavelength, the
  birefringence parameter used to calculate orientations, pixel scale, etc.
}
\description{
\code{KaminskyqPLM} creates an array-format qPLMarr object using
the standard output images from
\href{http://cad4.cpac.washington.edu/ROTOPOLhome/ROTOPOL.htm}{Rotopol}(tm)
software. User will be prompted to select a0, a1, and a2 images.
}
\details{
More information on the Rotopol(tm) system is available at
\url{http://cad4.cpac.washington.edu/ROTOPOLhome/ROTOPOL.htm}.
}
