% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBlock.R
\name{makeBlock}
\alias{makeBlock}
\title{Model an Embedded Tissue Block With Registration Marks}
\usage{
makeBlock(B, H, L1, S1, W1, Theta1 = 45, point1 = TRUE, L2, S2, W2,
  Theta2 = -45, point2 = TRUE)
}
\arguments{
\item{B}{Block width on face plane (micrometers?).}

\item{H}{Block height on face plane (micrometers?).}

\item{L1, L2}{Upper and lower longitudinal notch offsets. For each, this is
the distance from the nearest lateral face to the far side of the
longitudinal notch, as measured on the front or back faces of the block.}

\item{S1, S2}{Upper and lower oblique notch offsets. Measured from the nearest
lateral face to the *near* side of the oblique notch. Lx & Sx should be
measured on the same face (front or back).}

\item{W1, W2}{Upper and lower oblique notch widths, measured orthogonal to
oblique notch orientation (essentially the width of the notches as they
were machined).}

\item{Theta1, Theta2}{Angle between longitudinal and oblique notches upper
face and lower face. Default is 45 degrees.}

\item{point1, point2}{Do the upper/lower notches form an arrowhead pointing
toward the front face of the block?}
}
\value{
A object of type "histoBlock."
}
\description{
\code{makeBlock} creates a "histoBlock" object to register a
physical cross-section (typically a histology slide) within a MicroCT volume.
}
\seealso{
Other Section Registration Functions: \code{\link{registerSlide}}
}
