% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microTransit.R
\docType{package}
\name{microTransit}
\alias{microTransit}
\title{microTransit: R tools for Quantitative Polarized Light Microscopy (qPLM).}
\description{
\pkg{microTransit} provides tools for measuring, analyzing, and illustrating
3D orientation with transmitted polarized light. The tools in this package
were designed for the analysis of bone and connective tissue by microscopy on
thin sections, but the basic functions can be extended to assess orientation
in any birefringent material.
}
