% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullqPLMSample.R
\name{pullqPLMSample}
\alias{pullqPLMSample}
\title{Interactive Sub-sampling of qPLM Data by Pixel count}
\usage{
pullqPLMSample(qPLMarr, layer, pixels)
}
\arguments{
\item{qPLMarr}{A \code{qPLMarr} object.}

\item{layer}{Value from 1-3; this layer will be displayed on-screen for
selecting the ROI (1 = \strong{I}, 2 = \strong{theta}, 3 = \strong{phi}).
The resulting subsample will include all of the 'stacked' layers from the
\code{qPLMarr} object.}

\item{pixels}{approximate number of pixels required for the sample (rounding
for a square region may pull pixels more than specified by this argument).}
}
\value{
A \code{qPLMarr} object with the subsampled pixels. This object will
  maintain all the attributes of the original.
}
\description{
\code{pullqPLMSample} sub-samples a specified number of pixels
  from a \code{qPLMarr} object around an interactively selected point.
}
\examples{
#oldwd<-getwd()
#setwd(system.file("extdata", package = "microTransit"))
#load("testqPLMarr.R")
#setwd(oldwd)

testSample<-pullqPLMSample(testqPLMarr)

}
\seealso{
Other qPLM Analysis Functions: \code{\link{angGaussSumm}},
  \code{\link{centroidCorr}}, \code{\link{pullCluster}},
  \code{\link{pullROI}}, \code{\link{qPLMClust}},
  \code{\link{qPLMTabulate}}
}
