% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullROI.R
\name{pullROI}
\alias{pullROI}
\title{Interactive Sub-sampling of qPLM Data by Region of Interest}
\usage{
pullROI(qPLMarr, layer)
}
\arguments{
\item{qPLMarr}{A \code{qPLMarr} object.}

\item{layer}{Value from 1-3; this layer will be displayed on-screen for
selecting the ROI (1 = \strong{I}, 2 = \strong{theta}, 3 = \strong{phi}).
The resulting ROI will include all of the 'stacked' layers from the
\code{qPLMarr} object.}
}
\value{
A \code{qPLMarr} object with the subsampled pixels. This object will
  maintain all the attributes of the original.
}
\description{
\code{pullROI} sub-samples a \code{qPLMarr} object with an
  interactively selected rectangular ROI.
}
\examples{
#oldwd<-getwd()
#setwd(system.file("extdata", package = "microTransit"))
#load("testqPLMarr.R")
#setwd(oldwd)

testROI<-pullROI(testqPLMarr, 2)

}
\seealso{
Other qPLM Analysis Functions: 
\code{\link{angGaussSumm}()},
\code{\link{centroidCorr}()},
\code{\link{pullCluster}()},
\code{\link{pullqPLMSample}()},
\code{\link{qPLMClust}()},
\code{\link{qPLMTabulate}()}
}
\concept{qPLM Analysis Functions}
