% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPLMClust.R
\name{qPLMClust}
\alias{qPLMClust}
\title{Spatial Clustering of Axial Orientation Data from qPLM}
\usage{
qPLMClust(qPLM, cortical = FALSE, grainSize = 20, cutoff = 20,
  criterion = "theta", thresHold = 0.8, multi = TRUE)
}
\arguments{
\item{qPLM}{A \code{qPLMarr} or \code{qPLMtab} object.}

\item{grainSize}{Starting subsample dimensions in pixels. \code{qPLM.clust}
will cluster square blocks of \code{grainSize} x \code{grainSize} pixels
based on the angular central Gaussian distribution of their orientations.
Default is a 20 by 20 pixel area (400 pixels), which allows a balance
between discriminating smaller tissue features and providing enough
individual pixels for a tight estimate of distribution.}

\item{cutoff}{Minimum number of measured pixels present for a block to be
processed. If a \code{grainSize} x \code{grainSize} block has fewer than
\code{cutoff} measured pixels, it will be dropped. Intended to guard
against odd block parameters from edge effects.}

\item{criterion}{default is \code{"theta,"} which defines clusters based on
  the magnitude of the minor axis eigenvalue (low - high values reveal more
  or less anisotropy, respectively) and the out-of-plane angles of the major
  and second axes of the angular central Gaussian distribution of collagen
  fiber orientation per \code{grainSize} x \code{grainSize}) block. This is
  roughly equivalent to clustering by brightness in circularly polarized
  light microscopy, with some added nuance.

  Other available criteria:

  \code{"anisotropy"}, which defines clusters based solely on the eigenvalue
  magnitude of all three axes, and discards information on the orientation of
  the axes (experimental, may be useful for defining kinked/bent tendons or
  ligaments vs. surrounding tissue);

  \code{"custom"}, which allows users to select from the nine independent
  angular central Gaussian descriptors: \describe{ \item{[1:3]}{1st - 3rd
  axis eigenvalues;} \item{[4:6]}{1st axis Euclidean coordinates (x,y,z);}
  \item{[7:9]}{2nd axis Euclidean coordinates (x,y,z).} } To specify which
  descriptors to use, the user must create a list object named "customCrit"
  that contains the descriptor numbers as listed above. \code{qPLMClust} will
  search the local environment for \code{customCrit}.}

\item{thresHold}{\code{qPLMClust} will plot all groupings that have a
similarity score of less  than this variable. Useful for getting back to
major tissue differences when there are artifacts in the image that drown
out biologically relevant tissue dissimilarity.}

\item{multi}{Experimental. Estimating the angular central Gaussian
distribution in parallel using \code{foreach} is much faster,
but sometimes the array binding step using \code{abind} hits a length
mismatch and fails. Until the source of the error can be pinned down,
setting \code{multi} to \code{FALSE} reverts to a nested \code{for}
loop to work around the problem.}
}
\value{
Returns a qPLMclust object, which is a list with four components:
  \enumerate{ \item $qPLMtab: the pixel-by-pixel orientatin data in
  \code{qPLMtab} format. \item $GaussRaster: a \code{RasterBrick}
  representation of the angular central Gaussian distribution parameters per
  block. \item $groupMatrix: a matrix with rows that represent processed
  pixel blocks, and columns that represent successively more inclusive
  agglomerative spatial clusters. The cells in each column give the group
  membership for each block at that level, and the contents of the bottom
  cell in each column give the similarity value for that agglomerative step.
  \item $groupRaster: a \code{RasterBrick} representation of the clusters
  that merge at lower similarity values than the value specified by the
  \code{thresHold} variable. \code{pullCluster} calls on this component to
  let the user page through the high-level clusters and select a level that
  reflects their specific biological question (e.g., clusters that separate
  an area of tendon insertion relative to surrounding periosteal bone). }
}
\description{
\code{qPLMClust} produces a hierarchical set of groups that
  describe contiguous regions of a specimen with slow axis orientations that
  are more similar to each other than they are to neighboring regions.
}
\details{
\code{qPLMClust} uses iterative agglomerative clustering (Lance &
  Williams 1966; 1967) based on Euclidean distances between parameters of the
  angular central Gaussian distribution (Tyler 1987), constrained by spatial
  neighborhood (Legendre & Legendre 2012:839 - 844).

  This function provides an automated means of selecting areas of tissue that
  have a distinct "fabric" (e.g., mid-cortical vs. endosteal bone in long
  bone cross-section). It was specifically written to avoid the multiple
  arbitrary judgement calls that must be made when trying to sub-sample a
  bone cross-section using several small ROIs to stand in for a "class" of
  tissue.

  The comparison is blind to the biological significance of orientation
  differences--it is up to the user to select an appropriate hierarchical
  level of comparison.
}
\examples{
#oldwd<-getwd()
#setwd(system.file("extdata", package = "microTransit"))
#load("testqPLMarr.R")
#load("testqPLMtab.R")
#setwd(oldwd)

testqPLMclust_arr<-qPLMClust(testqPLMarr)
testqPLMclust_tab<-qPLMClust(testqPLMtab)


}
\references{
Lance, G.N., and Williams, W.T., 1966. A generalized sorting
  strategy for computer classifications. \emph{Nature 212}:218.

  Lance, G.N., and Williams, W.T., 1967. A general theory of classificatory
  sorting strategies. I. Hierachical systems. \emph{Computer J 9}:373 - 380.

  Legendre, P., and Legendre, L., 2012. \emph{Numerical Ecology}. Elsevier.

  Tyler, D.E., 1987. Statistical analysis for the angular central Gaussian
  distribution on the sphere. \emph{Biometrika, 74(3)}: 579 - 589.
}
\seealso{
Other qPLM Analysis Functions: \code{\link{angGaussSumm}},
  \code{\link{centroidCorr}}, \code{\link{pullCluster}},
  \code{\link{pullROI}}, \code{\link{pullqPLMSample}},
  \code{\link{qPLMTabulate}}
}
