% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPLMKeymaker.R
\name{qPLMKeymaker}
\alias{qPLMKeymaker}
\title{Create an Orientation Key for Interpreting qPLM Overview Images}
\usage{
qPLMKeymaker(pixel.size)
}
\arguments{
\item{pixel.size}{Dimension in pixels for square output image.}
}
\value{
Silent return in R. Saves an image with the name "x_pixel_key.tif"
  where x is the value of pixel.size, and displays the image to a browser
  window.
}
\description{
\code{qPLMKeymaker} creates a key image for interpreting
combined colatitude (theta) and longitude (phi) "_overview.tif" images.
}
\details{
The \code{.tiff} file produced by this function should appear as a
  colorful circle on a black background. Imagine this image as a view of the
  northern hemisphere of a globe, with the prime meridian directed to the right.
  A ray projecting from the center of the globe through the north pole
  (colatitude 0째) would emerge in the center of the circle, and be color-coded
  black. A ray emerging with a colatitude of 45째and a longitude of 45째would
  be color-coded as a half-saturated blue. A ray emerging at the equator with
  a longitude of 135째 would be color-coded as a fully saturated green.
}
\seealso{
Other qPLM Illustration Functions: 
\code{\link{acgEllipse}()},
\code{\link{qPLMSphereSkinner}()},
\code{\link{qPLMTiff}()}
}
\concept{qPLM Illustration Functions}
