% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPLMSphereSkinner.R
\name{qPLMSphereSkinner}
\alias{qPLMSphereSkinner}
\title{Create an 3D Skin for Interpreting qPLM Overview Images}
\usage{
qPLMSphereSkinner(pixel.size, theta.max)
}
\arguments{
\item{pixel.size}{Dimension in pixels for square output image.}

\item{theta.max}{Maximum value of theta. Used when the qPLM object to
interpret has a truncated theta range--take this value directly from the
"theta.max" attribute of the qPLM object.}
}
\value{
Silent return in R. Saves an image with the name "x_pixel_skin.tif"
  where x is the value of pixel.size, and displays the image to a browser
  window.
}
\description{
\code{qPLMSphereSkinner} creates a skin image for placing
colatitude (theta) and longitude (phi) on a sphere (e.g., in Adobe
Illustrator).
}
\details{
the \code{.tiff} image produced by this function can be used to skin
  a hemispherical object for inclusion in 3D-modeled figures.
}
\seealso{
Other qPLM Illustration Functions: 
\code{\link{acgEllipse}()},
\code{\link{qPLMKeymaker}()},
\code{\link{qPLMTiff}()}
}
\concept{qPLM Illustration Functions}
