% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPLMTabulate.R
\name{qPLMTabulate}
\alias{qPLMTabulate}
\title{Place qPLMarr Data in Tabular Format}
\usage{
qPLMTabulate(qPLMarr, low.pass = 5/256)
}
\arguments{
\item{qPLMarr}{A qPLMarr object.}

\item{low.pass}{A low-pass filter value (as an 8-bit grayscale level, 1-256)
to remove noisy in-plane orientation data from low-theta pixels. Default is
five.}
}
\value{
A [u*v,9] qPLMtab matrix, with columns (1) theta, (2) phi, (3-5) the
  Euclidean projection of theta and phi, (6-7) the [u,v] address of the pixel
  as an integer, and (8-9) the [u,v] position of the pixel in millimeters.
}
\description{
\code{qPLMTabulate} casts data from the [u,v,3] array format of
  a qPLMarr object into a qPLMtab matrix.
}
\examples{
#oldwd<-getwd()
#setwd(system.file("extdata", package = "microTransit"))
#load("testqPLMarr.R")
testqPLMtab<-qPLMTabulate(testqPLMarr)
testqPLMtab[1:5,]
save(testqPLMtab, file = "testqPLMtab.R")
#setwd(oldwd)

}
\seealso{
Other qPLM Analysis Functions: 
\code{\link{angGaussSumm}()},
\code{\link{centroidCorr}()},
\code{\link{pullCluster}()},
\code{\link{pullROI}()},
\code{\link{pullqPLMSample}()},
\code{\link{qPLMClust}()}
}
\concept{qPLM Analysis Functions}
