% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPLMTiff.R
\name{qPLMTiff}
\alias{qPLMTiff}
\title{Produce Overview TIFFs From qPLM Arrays}
\usage{
qPLMTiff(
  sample.name,
  qPLMarr,
  theta.scale = FALSE,
  theta.max = 90,
  invert.theta.scale = FALSE
)
}
\arguments{
\item{sample.name}{An identifier for the sample, used as a stem for the
output file names.}

\item{qPLMarr}{A qPLMarr object, e.g., output from \code{buildqPLM}.}

\item{theta.scale}{Boolean; if TRUE, automatically scales "_overview.tif"
image color scale to peak at maximum measured theta value (default is
FALSE).}

\item{theta.max}{Numerical; if theta.scale is FALSE, specifies an upper end
(degrees) for "_overview.tif" color scale (default is 90 degrees).}

\item{invert.theta.scale}{Boolean; if TRUE, bright colors point toward
viewer, brightness falls off as orientations go to parallel with the slide.
Useful for sections with predominantly out-of-plane fibers (default is
FALSE).}
}
\value{
Null; .tiff files saved to working directory.
}
\description{
\code{qPLMTiff} creates greyscale .tiff images of out-of-plane
  orientation (theta), in-plane orientation (phi), transmittance (trans), and
  a false-color representation of combined orientation data (overview).
}
\details{
The false-color image of the "overview" \code{.tiff} uses the
  LUV colorspace to produce a perceptually similar brightness for equal
  levels of theta, and encodes phi with different hues.
}
\seealso{
Other qPLM Illustration Functions: 
\code{\link{acgEllipse}()},
\code{\link{qPLMKeymaker}()},
\code{\link{qPLMSphereSkinner}()}
}
\concept{qPLM Illustration Functions}
