% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rotosim.R
\name{Rotosim}
\alias{Rotosim}
\title{Single-Pixel Simulation of Rotating Polarizer qPLM}
\usage{
Rotosim(steps = 6, sample)
}
\arguments{
\item{steps}{Integer; the number of analyzer positions to use. Default is
six.}

\item{sample}{a Jones matrix representation of a birefringent sample. Can be
made using \code{jones.ret}.}
}
\value{
A list containing (1) |sin d|, a measure of retardance (varies with
  out-of-plane angle), (2-3) phi, the in-plane orientation of the specimen
  slow axis, and (4) Io, the transmissivity of the specimen.
}
\description{
\code{Rotosim} uses Jones calculus to simulate the interactions
of a right circular polarizer, a birefringent specimen, and a rotating
analyzer in sequence, modeling the setup of a Glazer et al. (1996) method
qPLM scope.
}
\references{
Glazer, A.M., Lewis, J.G., and Kaminsky, W., 1996. An automatic
  optical imaging system for birefringent media. \emph{Proc R Soc A
  452(1955)}: 2751 - 2765.
}
\seealso{
Other qPLM Simulation Functions: \code{\link{jonesPol}},
  \code{\link{jonesRet}}
}
