% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densBiref.R
\name{densBiref}
\alias{densBiref}
\title{Estimate Birefringence of Bone Tissue From Mineral Density}
\usage{
densBiref(density, dHA = 3.16, birefHA = -0.004, birefC = 0.008933)
}
\arguments{
\item{density}{estimate of mineral density in the sample, in units of
mgHA/ccm.}

\item{dHA}{density of hydroxyapatite in SI units. Default is 3.16.}

\item{birefHA}{birefringence of hydroxyapatite. Default is -4e-3.}

\item{birefC}{birefringence of collagen. Default is 8.933e-3.}
}
\value{
numeric value giving the estimated birefringence.
}
\description{
\code{densBiref} uses a naive analytical approach to estimating
birefringence from bone mineral density, as expressed in "milligrams of
hydroxyapatite per cubic centimeter" (mgHA/ccm).
}
\seealso{
Other qPLM Input Functions: \code{\link{buildqPLM}}
}
