% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jonesPol.R
\name{jonesPol}
\alias{jonesPol}
\title{Jones Matrix Representation of a Linear Polarizer Element}
\usage{
jonesPol(phi)
}
\arguments{
\item{phi}{in-plane angle for the axis of the polarizer element.}
}
\value{
A Jones matrix.
}
\description{
\code{jonesPol} creates a Jones matrix linear polarizer element
at specified orientation phi.
}
\details{
none yet.
}
\seealso{
Other qPLM Simulation Functions: \code{\link{Rotosim}},
  \code{\link{jonesRet}}
}
