% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullCluster.R
\name{pullCluster}
\alias{pullCluster}
\title{Interactive Selection of Clustered Pixels}
\usage{
pullCluster(qPLMclust)
}
\arguments{
\item{qPLMclust}{A \code{qPLMclust} object, returned from \code{qPLMClust}.}
}
\description{
\code{pullCluster} sub-samples a spatially clustered qPLM
  dataset (a \code{qPLMclust} object) and returns a \code{qPLMtab} object
  with all the pixels in the selected cluster.
}
\details{
no.
}
\examples{
#oldwd<-getwd()
#setwd(system.file("extdata", package = "microTransit"))
#load("testqPLMclust.R")
#setwd(oldwd)

testCluster<-pullCluster(testqPLMclust_arr)

}
\seealso{
Other qPLM Analysis Functions: \code{\link{angGaussSumm}},
  \code{\link{centroidCorr}}, \code{\link{pullROI}},
  \code{\link{pullqPLMSample}}, \code{\link{qPLMClust}},
  \code{\link{qPLMTabulate}}
}
