% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerSlide.R
\name{registerSlide}
\alias{registerSlide}
\title{Register the Position of a Slide in a Notched Block}
\usage{
registerSlide(block, D1, F1, D2, F2)
}
\arguments{
\item{block}{A histoBlock object created by \code{make.block}.}

\item{D1, D2}{Distance on slide from far edge of longitudinal notch to near
edge of oblique notch, for upper/lower face, respectively.}

\item{F1, F2}{Distance on slide from near edge to far edge of the oblique
notch.}
}
\value{
A list specifying (1) the z-axis position of the slide's centroid
  "Z", and (2-4) an Euler rotation for the slide's orientation relative to
  the block's front face, specified as "Euler.X", "Euler.Y", and
  "Euler.Angle".
}
\description{
\code{registerSlide} uses notch measurements on a slide,
  together with the notch information from a block object (see
  \code{make.block}) to position an arbitrarily cut section within the
  reference frame of a machined embedded block. This allows precision
  registration of plastic-embedded ground sections to a volume representation
  of the block, e.g. a MicroCT data set.
}
\details{
forthcoming.
}
\seealso{
Other Section Registration Functions: \code{\link{makeBlock}}
}
